move rollbar to own module so wrapping can happen at top
